src/utils/hooks/usePagination.ts
type usePaginationProps = {
  totalPages: number;
  currentPage: number;
  setCurrentPage: (page: number) => void;
};

export default function usePagination({
  totalPages,
  currentPage,
  setCurrentPage,
}: usePaginationProps) {
  const handleNextPage = () => {
    if (currentPage < totalPages) {
      setCurrentPage(currentPage + 1);
    }
  };

  const handlePage = (pageNumber: number) => {
    setCurrentPage(pageNumber);
  };

  const handlePreviousPage = () => {
    if (currentPage > 1) {
      setCurrentPage(currentPage - 1);
    }
  };

  return {
    handleNextPage,
    handlePage,
    handlePreviousPage,
  };
}
